[primitives/node] rename api to api_ext
