+++ primitives/node/src/lib.rs
@@ -24,7 +24,7 @@ pub type ShieldFundsFn = ([u8; 2], Vec<u8>, u128, ShardIdentifier);
 pub type CallWorkerFn = ([u8; 2], Request);
 
 #[cfg(feature = "std")]
-pub mod api;
+pub mod api_ext;
 
 #[cfg(feature = "std")]
-pub use api::*;
+pub use api_ext::*;
